This is a good first functional milestone.
I'm able to record results in a time-series format for time-series type data, and machine-level format for status-type information. Going forward, the results SHOULD be posted to a Google Function, to make the system maximally functional with minimal need to revise the running clients. But for now, with appropriate login access to the clients, this will do.